pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "adafruit-blinka>=8.62.0",
    "adafruit-circuitpython-dht>=4.0.9",
    "apache-libcloud>=3.8.0",
    "click>=8.2.1",
    "miniupnpc>=2.3.3",
    "plotext>=5.3.2",
    "pyright>=1.1.403",
    "pyserial>=3.5",
    "pyserial-asyncio>=0.6",
    "ruff>=0.12.8",
    "textual>=5.3.0",
    "textual-dev>=1.7.0",
    "textual-plotext>=1.0.1",
]
